SceneHistoryUI : Adopt location drop from `startup/gui/graphEditor.py`
It's a natural fit alongside a bunch of other functionality based around `SceneAlgo.source()` and the like. And continues us towards the goal of most optional functionality being bolted on by configs calling `Foo.connectToEditor()`.
