+++ python/GafferSceneUI/SceneHistoryUI.py
@@ -35,6 +35,9 @@
 ##########################################################################
 
 import functools
+import weakref
+
+import imath
 
 import IECore
 import IECoreScene
@@ -69,6 +72,10 @@ def connectToEditor( editor ) :
 		editor.keyPressSignal().connect( __hierarchyViewKeyPress )
 	elif isinstance( editor, GafferUI.NodeEditor ) :
 		editor.keyPressSignal().connect( __nodeEditorKeyPress )
+	elif isinstance( editor, GafferUI.GraphEditor ) :
+		editor.graphGadget().dragEnterSignal().connect( __graphGadgetDragEnter )
+		editor.graphGadget().dragLeaveSignal().connect( __graphGadgetDragLeave )
+		editor.graphGadget().dropSignal().connect( functools.partial( __graphGadgetDrop, graphEditor = weakref.ref( editor ) ) )
 
 ##########################################################################
 # Internal implementation
@@ -220,3 +227,80 @@ def __nodeEditorKeyPress( nodeEditor, event ) :
 		if selectedPath is not None :
 			__editTweaksNode( nodeEditor.scriptNode().context(), scene, selectedPath, nodeEditor )
 		return True
+
+def __dropLocationData( event ) :
+
+	if (
+		not isinstance( event.data, IECore.StringVectorData ) or
+		len( event.data ) != 1 or
+		not event.data[0].startswith( "/" ) or
+		event.sourceWidget is None
+	) :
+		return None
+
+	scene = None
+	sourceEditor = event.sourceWidget.ancestor( GafferUI.Editor )
+	if isinstance( sourceEditor, GafferUI.Viewer ) :
+		if isinstance( event.sourceGadget, GafferSceneUI.SceneGadget ) :
+			scene = sourceEditor.view()["in"].getInput()
+	elif isinstance(
+		sourceEditor,
+		( GafferSceneUI.HierarchyView, GafferSceneUI.LightEditor, GafferSceneUI.AttributeEditor, GafferSceneUI.LightLinkingEditor )
+	) :
+		scene = sourceEditor.settings()["in"].getInput()
+
+	if scene is None :
+		return None
+
+	return {
+		"path" : event.data[0],
+		"scene" : scene,
+		"context" : sourceEditor.getContext(),
+	}
+
+def __graphGadgetDragEnter( graphGadget, event ) :
+
+	if __dropLocationData( event ) is None :
+		return False
+
+	GafferUI.Pointer.setCurrent( "targetObjects" )
+	return True
+
+def __graphGadgetDragLeave( graphGadget, event ) :
+
+	if __dropLocationData( event ) is not None :
+		if event.destinationWidget is None :
+			# Hack to restore (what we assume to have been) the original
+			# drag pointer. We don't do this when another widget has
+			# accepted the drag, because it would clobber any pointer
+			# change they made in `dragEnter`. But of course that means
+			# that if another widget _has_ accepted the drag but hasn't
+			# changed the pointer themselves (maybe they use highlighting
+			# instead), they will be stuck with the wrong pointer.
+			## \todo This is far too fragile. We need to manage
+			# pointer restoration at a higher level, in Widget.py's
+			# _EventFilter (and ViewportGadget's handlers).
+			GafferUI.Pointer.setCurrent( "objects" )
+		return True
+
+	return False
+
+def __graphGadgetDrop( graphGadget, event, graphEditor ) :
+
+	dropLocationData = __dropLocationData( event )
+	if dropLocationData is None :
+		return False
+
+	with dropLocationData["context"] :
+		sourceScene = GafferScene.SceneAlgo.source( dropLocationData["scene"], dropLocationData["path"] )
+
+	if sourceScene is not None :
+		graphGadget.setRoot( sourceScene.node().parent() )
+		## \todo The `frame()` method should probably be on the GraphGadget itself, and the `at`
+		# functionality should be made public.
+		graphEditor()._GraphEditor__frame(
+			[ sourceScene.node() ],
+			at = imath.V2f( GafferUI.Widget.mousePosition( relativeTo = graphEditor() ) )
+		)
+
+	return True
